# src/main/java/com/speer/notes/model/UserDto.java
package com.speer.notes.model;

public class UserDto {

	private String username;

	private String password;

	public String getUsername() {
		return username;
	}

	public void setUsername(String username) {
		this.username = username;
	}

	public String getPassword() {
		return password;
	}

	public void setPassword(String password) {
		this.password = password;
	}
}